Chapitre 3 terminé ! close #3

+++ TP3/src/Point.cpp
@@ -9,7 +9,7 @@ x(0),y(0),z(0)
 {}
 
 //constructeur par paramètre
-Point::Point(int x,int y,int z):
+Point::Point(float x,float y,float z):
 x(x),y(y),z(z)
 {}
